requirements.txt
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coloredlogs==15.0.1
flatbuffers==24.12.23
humanfriendly==10.0
idna==3.10
mpmath==1.3.0
numpy==2.2.1
onnxruntime==1.20.1
packaging==24.2
piper-phonemize==1.1.0
piper-tts==1.2.0
protobuf==5.29.2
pycparser==2.22
requests==2.32.3
sounddevice==0.5.1
sympy==1.13.3
urllib3==2.3.0
